+++ src/ansys/additive/core/parametric_study/parametric_study_progress_handler.py
@@ -56,6 +56,10 @@ class ParametricStudyProgressHandler(IProgressHandler):
         self._study_lock = threading.Lock()
         self._study = study
         self._additive = additive
+        # Store the last state of each simulation to avoid
+        # unnecessary disk writes when setting the simulation status
+        # on the study.
+        self._last_progress_states = {}
 
     def update(self, progress: Progress) -> None:
         """Update the progress of a simulation.
@@ -65,6 +69,13 @@ class ParametricStudyProgressHandler(IProgressHandler):
         progress : Progress
             Progress information for the simulation.
         """
+        if (
+            progress.sim_id in self._last_progress_states
+            and progress.state == self._last_progress_states[progress.sim_id]
+        ):
+            return
+
+        LOG.debug(f"Updating progress for {progress.sim_id}")
 
         if progress.state == ProgressState.WAITING:
             self._update_simulation_status(progress.sim_id, SimulationStatus.PENDING)
@@ -82,6 +93,8 @@ class ParametricStudyProgressHandler(IProgressHandler):
             self._update_simulation_results(progress.sim_id)
             self._update_simulation_status(progress.sim_id, SimulationStatus.COMPLETED)
 
+        self._last_progress_states[progress.sim_id] = progress.state
+
     def _update_simulation_status(
         self, sim_id: str, status: SimulationStatus, message: str = None
     ) -> None:
@@ -98,7 +111,7 @@ class ParametricStudyProgressHandler(IProgressHandler):
         """
 
         with self._study_lock:
-            self._study.set_status(sim_id, status, message)
+            self._study.set_simulation_status(sim_id, status, message)
 
     def _update_simulation_results(self, sim_id: str) -> None:
         """Update the results of a completed simulation.