+++ CHANGELOG.md
@@ -10,9 +10,11 @@
 * Default version of Additive Server is 25.1 when creating a client connection.
 * Single bead simulations with thermal history are now supported.
 * Added description field to AdditiveMaterial.
-* Updated the parametric runner to accept a list of simulation ids to be run.
+* Updated the parametric runner to run a list of simulation ids.
 * Asynchronous simulate functionality added.
 * Added apply and list server settings functionality.
+* Parametric study data is updated as simulations are running.
+* Use LOG for logging in Additive object.
 
 ### Bug Fixes